Added Cards to Contact Page
