+++ client-react/package.json
@@ -6,12 +6,14 @@
     "axios": "^0.19.2",
     "bootstrap": "^4.5.3",
     "jquery": "^3.5.1",
+    "mdbootstrap": "^4.19.1",
+    "mdbreact": "^4.27.0",
+    "popper.js": "^1.12.9",
     "react": "^16.13.1",
     "react-bootstrap": "^1.3.0",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "3.0.1",
-    "popper.js": "^1.12.9"
+    "react-scripts": "3.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
