+++ client-react/src/App.js
@@ -1,17 +1,34 @@
 import React from "react";
-import  Ticket from "./components/Home";
-import "./components/App.css";
-import 'bootstrap/dist/css/bootstrap.min.css';
+import Home from "./components/Home";
 import Navbar from "./components/Navbar/Navbar";
+import ViewTicket from "./components/ViewTicket";
+
+
+import Cards from "./components/Cards/Card"
+
+import "./App.css";
+import { BrowserRouter as Router, Route} from 'react-router-dom';
 
 function App() {
   return (
-    <div className="App">
-      <Navbar />
-      <Ticket />
+    <Router>
+      <div>
+        <Navbar />
 
-    </div>
+        
+
+        <Route exact path="/" component={Home} />
+        <Route path="/view" component={ViewTicket} />
+        <Route path="/contact" component={Cards} />
+        
+
+      </div>
+    </Router>
   );
 }
+
+
+
+
 
 export default App;
