+++ client-react/src/components/Button.js
@@ -7,6 +7,7 @@ const STYLES = [
 ]
 
 const SIZES = [
+    
     'btn--medium',
     'btn--large'
 ]
