+++ client-react/src/components/Cards/CardUi.js
@@ -1,0 +1,25 @@
+import React from 'react'
+
+import "./cardStyle.css";
+
+
+const Card = props => {
+
+
+return (
+    <div className="card text-center">
+        <div className="overflow">
+            <img src= {props.imgsrc} alt="Image 1" className="card-img-top" />
+        </div>
+    <div className="card-body text-dark">
+    <h4 className="card-title">{props.title}</h4>
+    <p className="card-text text-secondary">blah blah blah </p>
+    <a href="#" className="btn btn-outline-success">{props.link}</a>
+    </div>
+    </div>
+
+)
+
+}
+
+export default Card;
