+++ client-react/src/components/Cards/cardStyle.css
@@ -1,0 +1,38 @@
+body {
+    background: radial-gradient(#e5e5e5, #fff, #e5e5e5);
+}
+
+.card {
+    width: 20rem;
+
+}
+
+.card:hover{
+    box-shadow: 5px 10px 20px 1px rgba(0, 0, 0, 0.253)!important;
+}
+
+.card-body {
+    padding: 3rem !important;
+}
+
+.card-text {
+    font-size:0.9rem;
+    padding: 0.4rem 1.9rem;
+}
+
+.container-fluid .row {
+    padding: 6rem;
+}
+
+.overflow{
+    
+}
+
+.card-img-top {
+    transform: scale(1);
+    transition: transform 0.5s ease;
+}
+
+.card-img-top:hover{
+    transform: scale(1.8);
+}
