+++ client-react/src/components/Navbar/Navbar.css
@@ -1,6 +1,7 @@
 .NavbarItems {
     background: linear-gradient(90deg, rgb(110,94,254) 0%, rgba(73,63,252,1) 100%);
     height: 80px;
+    box-shadow: 3px 3px 3px 3px grey;
     display: flex;
     align-items: center;
     font-size: 1.2rem;
@@ -57,6 +58,7 @@
 @media screen and (max-width: 480px) {
     .NavbarItems {
         position: relative;
+        box-shadow: 3px 3px 3px 3px grey;
     }
 
 .nav-menu {
@@ -68,14 +70,14 @@
     top: 80px;
     left: -100%;
     opacity: 1;
-    transition: all 0.5 ease;
+    transition: all 0.5s ease;
     }
 
 .nav-menu.active {
     background: #6668f4;
     left: 0;
     opacity: 1;
-    transition: all 0.5 ease;
+    transition: all 0.5s ease;
     z-index: 1;
 }
 
@@ -144,6 +146,7 @@ Button {
 @media screen and (max-width: 768px) {
     .NavbarItems {
         position: relative;
+        box-shadow: 3px 3px 3px 3px grey;
     }
 
 .nav-menu {
@@ -231,6 +234,8 @@ Button {
 @media screen and (max-width: 1024px) {
     .NavbarItems {
         position: relative;
+        box-shadow: 3px 3px 3px 3px grey;
+        
     }
 
 .nav-menu {
@@ -242,14 +247,14 @@ Button {
     top: 80px;
     left: -100%;
     opacity: 1;
-    transition: all 0.5 ease;
+    transition: all 0.5s ease;
     }
 
 .nav-menu.active {
     background: #6668f4;
     left: 0;
     opacity: 1;
-    transition: all 0.5 ease;
+    transition: all 0.5s ease;
     z-index: 1;
 }
 
@@ -309,8 +314,8 @@ Button {
     color: #6568F4;
     transition: 250ms;
 }
-Button {
+/* Button {
     display: none !Important;
-}
+} */
 
 }
