+++ client-react/src/index.js
@@ -1,12 +1,16 @@
 import 'bootstrap/dist/css/bootstrap.min.css';
-import $ from 'jquery';
-import Popper from 'popper.js';
+
 import 'bootstrap/dist/js/bootstrap.bundle.min';
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './components/index.css';
 import App from './App'
 import * as serviceWorker from './serviceWorker';
+import "../node_modules/bootstrap/dist/css/bootstrap.min.css";
+import '@fortawesome/fontawesome-free/css/all.min.css'; 
+import'bootstrap-css-only/css/bootstrap.min.css';
+ import'mdbreact/dist/css/mdb.css';
+
 
 ReactDOM.render(<App />, document.getElementById('root'));
