+++ server-java-mysql-2/src/main/java/com/example/groupproject/models/Ticket.java
@@ -1,0 +1,116 @@
+package com.example.groupproject.models;
+
+//import java.sql.Timestamp;
+import java.util.Date;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+
+import org.hibernate.annotations.CreationTimestamp;
+
+
+@Entity
+@Table(name = "ticket")
+public class Ticket {
+	
+	
+//	Identify all columns and data types
+	
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+	
+	@Column(columnDefinition = "boolean default false")
+	  private boolean complete;
+	
+	@CreationTimestamp
+	private Date timeStamp;
+	
+	private String firstName;	
+	private String lastName;
+	private String email;
+	private String content;
+	private String category;
+	private String urgency;
+	
+	
+	
+//	All Getters and Setters
+	
+	public Long getId() {
+		return id;
+	}
+	public void setId(Long id) {
+		this.id = id;
+	}
+	
+	
+	public boolean getComplete() {
+	    return this.complete;
+	}	
+	public void setComplete(boolean complete) {
+	    this.complete = complete;
+	}
+	
+	
+	public Date getTimeStamp() {
+		return timeStamp;
+	}
+	public void setTimeStamp(Date timeStamp) {
+//		Timestamp settimestamp = new Timestamp(System.currentTimeMillis());
+		this.timeStamp = timeStamp;
+	}
+	
+	
+	public String getFirstName() {
+		return firstName;
+	}
+	public void setFirstName(String firstName) {
+		this.firstName = firstName;
+	}
+	
+	
+	public String getLastName() {
+		return lastName;
+	}
+	public void setLastName(String lastName) {
+		this.lastName = lastName;
+	}
+	
+	
+	public String getEmail() {
+		return email;
+	}
+	public void setEmail(String email) {
+		this.email = email;
+	}
+
+	
+	public String getContent() {
+		return content;
+	}
+	public void setContent(String content) {
+		this.content = content;
+	}
+		  	
+	
+	public String getCategory() {
+		return category;
+	}
+	public void setCategory(String category) {
+		this.category = category;
+	}
+	
+	
+	public String getUrgency() {
+		return urgency;
+	}
+	public void setUrgency(String urgency) {
+		this.urgency = urgency;
+	}
+	
+}
