+++ server-java-mysql-2/src/main/java/com/example/groupproject/models/TicketRepository.java
@@ -1,0 +1,7 @@
+package com.example.groupproject.models;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TicketRepository extends JpaRepository<Ticket, Long> {
+
+}
