+++ server-java-mysql-2/src/main/resources/application.properties
@@ -1,0 +1,9 @@
+
+spring.datasource.url = jdbc:mysql://localhost:3306/finalproject?createDatabaseIfNotExist=true
+spring.datasource.username = root
+spring.datasource.password = Password1!
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
+
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.hibernate.ddl-auto = create
