+++ server-java-mysql-2/src/test/java/com/example/groupproject/GroupprojectApplicationTests.java
@@ -1,0 +1,16 @@
+package com.example.groupproject;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class GroupprojectApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
