--- server-java-mysql/src/main/java/com/example/groupproject/GroupProjectApplication.java
@@ -1,26 +1,0 @@
-package com.example.groupproject;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@SpringBootApplication
-public class GroupProjectApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(GroupProjectApplication.class, args);
-	}
-	
-	@Bean
-    public WebMvcConfigurer corsConfigurer() {
-        return new WebMvcConfigurer() {
-            @Override
-            public void addCorsMappings(CorsRegistry registry) {
-                registry.addMapping("/**").allowedOrigins("http://localhost:3000","http://localhost:4200");
-            }
-        };
-    }
-
-}
