--- server-java-mysql/src/main/java/com/example/groupproject/models/Task.java
@@ -1,45 +1,0 @@
-package com.example.groupproject.models;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name = "task")
-public class Task {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.AUTO)
-  private Long id;
-  private String name;
-  @Column(columnDefinition = "boolean default false")
-  private boolean complete;
-
-  public Long getId() {
-    return this.id;
-  }
-
-  public void setId(Long id) {
-    this.id = id;
-  }
-
-  public String getName() {
-    return this.name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-
-  public boolean getComplete() {
-    return this.complete;
-  }
-
-  public void setComplete(boolean complete) {
-    this.complete = complete;
-  }
-
-}
