--- server-java-mysql/src/main/java/com/example/groupproject/models/TaskRepository.java
@@ -1,7 +1,0 @@
-package com.example.groupproject.models;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TaskRepository extends JpaRepository<Task, Long> {
-
-}
