--- server-java-mysql/src/main/java/com/example/groupproject/models/TicketRepository.java
@@ -1,7 +1,0 @@
-package com.example.groupproject.models;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TicketRepository extends JpaRepository<Ticket, Long> {
-
-}
