--- server-java-mysql/src/main/resources/application.properties
@@ -1,13 +1,0 @@
-
-
-
-
-spring.datasource.url = jdbc:mysql://localhost:3306/groupprojects?createDatabaseIfNotExist=true
-
-spring.datasource.username = root
-spring.datasource.password = Password1!
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL5Dialect
-
-spring.jpa.show-sql=true
-spring.jpa.properties.hibernate.format_sql=true
-spring.jpa.hibernate.ddl-auto = create
